refactor: mvc 프레임워크와 애플리네이션 클래스 분리
