+++ nextstep-mvc/src/main/java/nextstep/mvc/DispatcherServlet.java
@@ -1,5 +1,6 @@
 package nextstep.mvc;
 
+import nextstep.mvc.tobe.AnnotationHandlerMapping;
 import nextstep.mvc.tobe.JspView;
 import nextstep.mvc.tobe.ModelAndView;
 import org.slf4j.Logger;
@@ -28,10 +29,23 @@ public class DispatcherServlet extends HttpServlet {
 
     @Override
     public void init() throws ServletException {
+        // TODO: 현재는 DispatcherServlet이 생성자로 HandlerMapping과 HandlerAdapter를 DI 받는데, 이가 null일때 디폴트를 처리해보자.
+        initDefaultHandlerMappings();
+        initDefaultHandlerAdapters();
         for (HandlerMapping handlerMapping : handlerMappings) {
             handlerMapping.initialize();
         }
     }
+
+    private void initDefaultHandlerMappings() {
+        handlerMappings.add(new AnnotationHandlerMapping("slipp.controller"));
+    }
+
+    private void initDefaultHandlerAdapters() {
+        handlerAdapters.add(new AnnotationHandlerAdapter());
+    }
+
+
 
     @Override
     protected void service(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
